edit code v.2

+++ app/Http/Controllers/userController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\user;
+use App\Models\borrow;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Carbon\Carbon;
@@ -42,13 +43,22 @@ class userController extends Controller
     public function delete($id)
     {
         try {
+
             // $user = User::findOrFail($id); //where find return ในกรณี อย่าส่ง 500 *handle
             // $user->delete();
-            $user = User::where('id', $id)->whereDoesntHave('borrows', function ($query) {
-                $query->where('borrow_status', 'borrowing');
-            })->first();
-            $user->delete();
-            return response()->json(['success' => true, 'message' => 'User deleted successfully'], 201); //success message error
+            // $user = User::where('id', $id)->whereDoesntHave('borrows', function ($query) {
+            //     $query->where('borrow_status', 'borrowing');
+            // })->first();
+            $user = User::withTrashed()->where('id', $id)->first();
+            if ($user) {
+                if (Borrow::where('user_id', $id)->where('borrow_status', '!=', 'คืนแล้ว')->exists()) {
+                    return response()->json(['success' => false, 'message' => 'User must return product before deletion'], 400);
+                }
+                $user->delete();
+                return response()->json(['success' => true, 'message' => 'User deleted successfully'], 201);
+            } else {
+                return response()->json(['success' => false, 'error' => 'User not found'], 404);
+            } //success message error
         } catch (\Exception $e) {
             return response()->json(['success' => false, 'message' => $e->getMessage()], 400);
         }
@@ -69,38 +79,46 @@ class userController extends Controller
                 $user->update($validated_data);
                 return response()->json(['success' => true, 'message' => 'User updated successfully'], 201);
             } else {
-                return response()->json(['success' => false,'error' => 'User not found'], 404);
+                return response()->json(['success' => false, 'error' => 'User not found'], 404);
             }
         } catch (\Exception $e) {
-            return response()->json(['success' => false,'error' => $e->getMessage()], 400);
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 400);
         }
     }
 
     public function getList(Request $request)
     {
         try {
-            $gender = $request->input('gender'); //ใช้ request สำหรับ update บางตัวได้ *enum เช่น เพศ status (มีขอบเขต)
+            //ใช้ request สำหรับ update บางตัวได้ *enum เช่น เพศ status (มีขอบเขต)
+            // $gender = $request->input('gender');
+            // if ($gender) {
+            //     $query->where('gender', $gender);
+            // }
             $query = User::query(); //user all
-            if ($gender) {
-                $query->where('gender', $gender);
+            $filters = $request->only(['name', 'times_of_borrow']);
+            foreach ($filters as $key => $value) {
+                $query->where($key, 'like', "%$value%");
             }
             $user_data = $query->get();
-            $user_data = $user_data->map(function ($user) { //ใช้ชื่อตาม $user_data หรือ ประกาศ record=[] ก่อน *check ก่อน map เสมอ
+            $records = [];
+            $records = $user_data->map(function ($user) { //ใช้ชื่อตาม $user_data หรือ ประกาศ record=[] ก่อน *check ก่อน map เสมอ
                 return [
                     'id' => $user->id,
                     'name' => $user->name,
+                    'gender' => $user->gender,
                     'times_of_borrow' => $user->times_of_borrow
                 ];
             });
-            return response()->json(['List of users' => $user_data]);
+            return response()->json(['success' => true, 'List of users' => $records], 200);
+
         } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 400);
         }
     }
     public function getDetail($id)
     {
         try {
-            $user = User::findOrFail($id); //find = user only no relation, where = condition relation
+            $user = User::find($id); //find = user only no relation, where = condition relation
             $age = $this->calculateAge($user->birthday);
             $record = [ //key
                 'id' => $user->id,
@@ -112,9 +130,9 @@ class userController extends Controller
                 'email' => $user->email,
                 'times_of_borrow' => $user->times_of_borrow
             ];
-            return response()->json(['Detail of User' => $record]); //
+            return response()->json(['success' => true, 'Detail of User' => $record], 200); //
         } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500); //
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 400); //
         }
     }
